AsynRunner Annotation mybatis interceptor

+++ src/main/java/com/pengjinfei/maven/service/integration/ServiceActivatorDemo.java
@@ -1,8 +1,10 @@
 package com.pengjinfei.maven.service.integration;
 
-import com.pengjinfei.maven.dto.Person;
+import com.pengjinfei.maven.entity.Person;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+
+import java.math.BigDecimal;
 
 /**
  * Created on 8/20/17
@@ -15,9 +17,9 @@ public class ServiceActivatorDemo {
 
     public void sayHello(Person person) {
         log.info("Hello! {}, welcome you!", person);
-        Integer age = person.getAge();
+        BigDecimal age = person.getAge();
         if (age != null) {
-            if (age % 2 != 0) {
+            if (age.intValue() % 2 != 0) {
                 throw new RuntimeException("test");
             }
         }
